[fix/#41]: 타입 변경

+++ src/main/java/fisa/woorizip/backend/consumptionanalysis/domain/ConsumptionAnalysis.java
@@ -37,5 +37,5 @@ public class ConsumptionAnalysis {
     private double grocery;
 
     @Column(name = "customer_count", nullable = false)
-    private int customerCount;
+    private long customerCount;
 }
